third_party/WebKit/LayoutTests/fast/dom/javascript-url-exception-isolation-expected.txt

CONSOLE ERROR: line 1: Uncaught 42
CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token <
Exceptions thrown in javascript URLs should not propagate to the main script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS caughtException is false
PASS caughtException is false
PASS successfullyParsed is true

TEST COMPLETE
